unreal_asset: fix compilation, fix clippy warnings

+++ unreal_asset/src/properties/movies/movie_scene_evaluation_key_property.rs
@@ -83,6 +83,6 @@ impl PropertyTrait for MovieSceneEvaluationKeyProperty {
 
         self.value.write(asset)?;
 
-        Ok((asset.position() - begin))
+        Ok((asset.position() - begin) as usize)
     }
 }
